requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
Flask-Limiter==3.8.0
cryptography==44.0.0
Werkzeug==3.1.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
Pillow==12.0.0
requests==2.32.5
python-dotenv==1.0.1
gunicorn==23.0.0